Tweak for Low-Memory usage. Make compatible with raspberry pi.
The orientDB tries to allocate too much memory for a small raspberry pi.
This commit reduces the RAM usage of the internally used orientdb.

Important: Start the docker container with
    --env ORIENTDB_OPTS_MEMORY="-Xms200M -Xmx200M -XX:MaxDirectMemorySize=200M"
to further reduce memory usage.

The settings for the orientdb have been found by try and error.
This is a list of all [db parameters](http://www.orientdb.com/docs/last/admin/Configuration.html).
I used some suggestions from
[this](https://stackoverflow.com/questions/37016787/orientdb-2-1-11-java-process-consuming-too-much-memory/46023793)
stack overflow question and just went through all the config parameters of the documentation
and tweaked those that sounded like they would affect memory consumption.

The original base image did not have any binaries for the arm32 architecture on docker hub.
In fact, I did not found a single arm32 image with openjdk 12 jre/jdk. (why?)
So I tried openjdk 14 and it seems to work.

The new image is larger and based on ubuntu, but that's better than nothing.

In the orientdb server.sh there is this piece of code:

    # Raspberry Pi check (Java VM does not run with -server argument on ARMv6)
    if [ `uname -m` != "armv6l" ]; then
      JAVA_OPTS="$JAVA_OPTS -server "
    fi

I don't know if this workaround is still needed for the architecture `armv7l` of
the newer Raspberry Pis, but I simply replaced armv6l by armv7l. Seems to work.